Borrow a column from another result set on the table and boxplot pages
A "borrow a column" control on both pages pulls one framework's results
in from another run (or experiment) for side-by-side comparison. The
choice lives in the from/col query params, so the URL is shareable; the
other set is only fetched once a source is picked.

The borrowed table column is set off by a dashed hairline and a
"borrowed" tag, and its header names the source run (with its CPU
throttle called out when it differs from the viewed run). The borrowed
boxplot row labels itself with the framework and the source run's name.
